+++ services/post/src/domain/post.model.ts
@@ -7,8 +7,6 @@ import { CreatePostDto } from './dto/post.dto';
 
 import crypto from 'crypto';
 
-/* model for created posts */
-
 export type CreatePostModelInput = {
   id: bigint;
   slug: string;
@@ -110,56 +108,5 @@ export class Post implements IPost {
     if (!this.likers.includes(user.getID())) throw new Error('not liked');
     this.likers = this.likers.filter((liker) => liker !== user.getID());
     this.likenum--;
-  }
-}
-
-/* model for posts waiting to be created */
-interface IPrePost {
-  getCreatePostForm(): CreatePostDto;
-}
-
-export type CreatePrePostModelInput = {
-  author: User;
-  images: string[];
-  content: string;
-};
-
-@Injectable()
-export class PrePost implements IPrePost {
-  private readonly author: User;
-  private readonly images: string[];
-  private readonly content: string;
-
-  private readonly slug: string;
-  private readonly likenum: number;
-
-  constructor(post: CreatePrePostModelInput) {
-    this.author = post.author;
-    this.images = post.images;
-    this.content = post.content;
-    this.slug = this.getSlug(post.content);
-    this.likenum = 0;
-  }
-
-  public getCreatePostForm(): CreatePostDto {
-    return {
-      author: this.author.getID(),
-      srcs: [...this.images],
-      content: this.content,
-      slug: this.slug,
-      likenum: this.likenum,
-    };
-  }
-
-  private getSlug(content: string): string {
-    let slug: string;
-    const slugEnd: number = content.indexOf('\n');
-    if (slugEnd == -1 || slugEnd > 30) {
-      slug = content.slice(0, 30);
-    } else {
-      slug = content.slice(0, slugEnd);
-    }
-    slug = encodeURIComponent(slug + crypto.randomBytes(8).toString('hex'));
-    return slug;
   }
 }
