+++ services/post/src/domain/preComment.model.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@nestjs/common';
+
+export type CreatePreCommentInput = {};
+@Injectable()
+export class PreComment {
+  private author: bigint;
+
+  private comments: Comment[] = [];
+
+  constructor() {}
+}
